requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp==1.0.5            # via discord.py, pyanimelist
async-timeout==1.4.0      # via aiohttp
beautifulsoup4==4.6.0
bs4==0.0.1                # via pyanimelist
chardet==3.0.4            # via aiohttp
dicttoxml==1.7.4          # via pyanimelist
discord.py==0.16.7
discord==0.0.2
feedparser==5.2.1
future==0.16.0            # via google-search-api
google-search-api==1.1.14
inflect==0.2.5
lxml==4.0.0               # via pyanimelist
mtgsdk==1.3.1
multidict==3.2.0          # via aiohttp, yarl
protobuf==3.4.0
pyanimelist==1.1.0
pyyaml==3.12              # via vcrpy
requests==2.14.1
selenium==2.53.6          # via google-search-api
six==1.11.0               # via protobuf, vcrpy
strawpy==1.1
unidecode==0.4.21         # via google-search-api
vcrpy==1.11.1             # via google-search-api
websockets==3.4           # via discord.py
wrapt==1.10.11            # via vcrpy
xkcd==2.4.2
yarl==0.14.0              # via vcrpy
